# [RFC] improve(core): Upgrade @openzeppelin/contracts from 3.0.0 to 3.3.0