--- a/packages/core/contracts/oracle/implementation/VotingToken.sol
+++ b/packages/core/contracts/oracle/implementation/VotingToken.sol
@@ -22,23 +22,15 @@ contract VotingToken is ExpandedERC20, ERC20Snapshot {
         return _snapshot();
     }
 
-    // _transfer, _mint and _burn are ERC20 internal methods that are overridden by ERC20Snapshot,
-    // therefore the compiler will complain that VotingToken must override these methods
-    // because the two base classes (ERC20 and ERC20Snapshot) both define the same functions
+    // _beforeTokenTransfer is an ERC20 internal method that is overridden by ERC20Snapshot,
+    // therefore the compiler will complain that VotingToken must override this method
+    // because the two base classes (ERC20 and ERC20Snapshot) both define the same function
 
-    function _transfer(
+    function _beforeTokenTransfer(
         address from,
         address to,
-        uint256 value
-    ) internal override(ERC20, ERC20Snapshot) {
-        super._transfer(from, to, value);
-    }
-
-    function _mint(address account, uint256 value) internal override(ERC20, ERC20Snapshot) {
-        super._mint(account, value);
-    }
-
-    function _burn(address account, uint256 value) internal override(ERC20, ERC20Snapshot) {
-        super._burn(account, value);
+        uint256 amount
+    ) internal virtual override(ERC20, ERC20Snapshot) {
+        super._beforeTokenTransfer(from, to, amount);
     }
 }