**Motivation**

*This is more of a RFC (request for comments) than an actual pull request, though getting this PR merged in some form as a result of the fruitful discussion, would certainly be considered a happy ending.*

We at @jarvis-network are facing issues with getting all of our contracts verified on Etherscan. After extensive debugging and investigation me and @Aleione noticed a difference in the bytecode between the published artifacts in the npm package of our fork of UMAProtocol and when building the same contracts from our project, even the though the compiler version and optimization settings are the same. After more investigation we narrowed down the problem to the version of openzeppelin used. In our case we have `@openzeppelin/contracts@3.3.0` and the core package is using `@openzeppelin/contracts@3.0.0 `. Downgrading is not an option for us, since we use features released in version `3.3.0`.
To make sure that compiling UMA contracts both from this repo (well actually our fork) and our project that includes it as an npm dependency yields the same bytecode (when compiled with the same solc version and same optimization settings), we decided to upgrade `@openzeppelin/contracts` to `3.3.0`. Getting the code to compile was simple enough - I just had to address one [breaking change](https://github.com/OpenZeppelin/openzeppelin-contracts/blob/master/CHANGELOG.md#breaking-changes) introduced by `@openzeppelin/contracts` `3.1.0` (yeah, so much for having semantic versioning...), however we immediately faced an issue with `hardhat test`. It seems like the upgrade of `@openzeppelin/contracts` alone is enough to tip over the deployment over the limit:

<details><summary>Click to expand the output of <code>hardhat test</code>:</summary>

```sh
➤ cd packages/core
➤ node --max-old-space-size=1024 (yarn bin hardhat) test --verbose
...
hardhat:core:hre Creating HardhatRuntimeEnvironment +0ms
...
...
hardhat:core:hre Running task compile:solidity:log:run-compiler-start +0ms
Compiling 105 files with 0.6.12
  hardhat:core:hre Running task compile:solidity:solc:run +1ms
...
...
Compilation finished successfully
...
...
Contract: Lockable
  hardhat:core:hre Running task run-truffle-fixture +19s
  hardhat:core:hre Creating provider for network hardhat +6ms

    1) "before all" hook: Running truffle fixture if available in "Contract: Lockable"

  Contract: MultiRole
  hardhat:core:hre Running task run-truffle-fixture +4s

    2) "before all" hook: Running truffle fixture if available for "Exclusive Self-managed role"
...
...
  0 passing (59s)
  36 failing

  1) Contract: Lockable
       "before all" hook: Running truffle fixture if available in "Contract: Lockable":
     Error: Transaction reverted: trying to deploy a contract whose code is too large
      at ExpiringMultiPartyLib.constructor (contracts/financial-templates/expiring-multiparty/ExpiringMultiPartyLib.sol:11)
      at $USER/code/repos/blockchain/eth/UMAprotocol/node_modules/@nomiclabs/truffle-contract/lib/execute.js:229:26
      at Function.new ($USER/code/repos/blockchain/eth/UMAprotocol/node_modules/@nomiclabs/truffle-contract/lib/contract/constructorMethods.js:58:53)
      at deploy ($USER/code/repos/blockchain/eth/UMAprotocol/packages/common/src/MigrationUtils.js:94:46)
      at module.exports (migrations/15_deploy_expiring_multi_party_creator.js:51:40)
      at runMicrotasks (<anonymous>)
      at processTicksAndRejections (internal/process/task_queues.js:97:5)
      at module.exports (test/truffle-fixture.js:34:3)
      at SimpleTaskDefinition.action ($USER/code/repos/blockchain/eth/UMAprotocol/node_modules/@nomiclabs/hardhat-truffle5/src/index.ts:185:5)
      at Environment._runTaskDefinition ($USER/code/repos/blockchain/eth/UMAprotocol/node_modules/hardhat/src/internal/core/runtime-environment.ts:217:14)
      at Environment.run ($USER/code/repos/blockchain/eth/UMAprotocol/node_modules/hardhat/src/internal/core/runtime-environment.ts:130:14)
      at Context.<anonymous> ($USER/code/repos/blockchain/eth/UMAprotocol/node_modules/@nomiclabs/hardhat-truffle5/src/index.ts:145:9)

  2) Contract: MultiRole
       "before all" hook: Running truffle fixture if available for "Exclusive Self-managed role":
     Error: Transaction reverted: trying to deploy a contract whose code is too large
      at ExpiringMultiPartyLib.constructor (contracts/financial-templates/expiring-multiparty/ExpiringMultiPartyLib.sol:11)
      at $USER/code/repos/blockchain/eth/UMAprotocol/node_modules/@nomiclabs/truffle-contract/lib/execute.js:229:26
      at Function.new ($USER/code/repos/blockchain/eth/UMAprotocol/node_modules/@nomiclabs/truffle-contract/lib/contract/constructorMethods.js:58:53)
      at deploy ($USER/code/repos/blockchain/eth/UMAprotocol/packages/common/src/MigrationUtils.js:94:46)
      at module.exports (migrations/15_deploy_expiring_multi_party_creator.js:51:40)
      at runMicrotasks (<anonymous>)
      at processTicksAndRejections (internal/process/task_queues.js:97:5)
      at module.exports (test/truffle-fixture.js:34:3)
      at SimpleTaskDefinition.action ($USER/code/repos/blockchain/eth/UMAprotocol/node_modules/@nomiclabs/hardhat-truffle5/src/index.ts:185:5)
      at Environment._runTaskDefinition ($USER/code/repos/blockchain/eth/UMAprotocol/node_modules/hardhat/src/internal/core/runtime-environment.ts:217:14)
      at Environment.run ($USER/code/repos/blockchain/eth/UMAprotocol/node_modules/hardhat/src/internal/core/runtime-environment.ts:130:14)
      at Context.<anonymous> ($USER/code/repos/blockchain/eth/UMAprotocol/node_modules/@nomiclabs/hardhat-truffle5/src/index.ts:145:9)
...
...
```

</details>

And `truffle test` yields the following:

<details><summary>Click to expand the output of <code>truffle test</code>:</summary>


```
➤ yarn truffle test
$ $USER/code/repos/blockchain/eth/UMAprotocol/node_modules/.bin/truffle test
Using network 'test'.


Compiling your contracts...
===========================
> Compiling ./contracts/Migrations.sol
> Compiling ./contracts/common/implementation/AddressWhitelist.sol
> Compiling ./contracts/common/implementation/ExpandedERC20.sol
> Compiling ./contracts/common/implementation/FixedPoint.sol
> Compiling ./contracts/common/implementation/Lockable.sol
> Compiling ./contracts/common/implementation/MultiRole.sol
> Compiling ./contracts/common/implementation/Testable.sol
> Compiling ./contracts/common/implementation/TestnetERC20.sol
> Compiling ./contracts/common/implementation/Timer.sol
> Compiling ./contracts/common/implementation/Withdrawable.sol
> Compiling ./contracts/common/interfaces/Balancer.sol
> Compiling ./contracts/common/interfaces/ExpandedIERC20.sol
> Compiling ./contracts/common/interfaces/IERC20Standard.sol
> Compiling ./contracts/common/interfaces/OneSplit.sol
> Compiling ./contracts/common/interfaces/TransactionBatcher.sol
> Compiling ./contracts/common/interfaces/Uniswap.sol
> Compiling ./contracts/common/test/BalancerMock.sol
> Compiling ./contracts/common/test/BasicERC20.sol
> Compiling ./contracts/common/test/MultiRoleTest.sol
> Compiling ./contracts/common/test/OneSplitMock.sol
> Compiling ./contracts/common/test/ReentrancyAttack.sol
> Compiling ./contracts/common/test/ReentrancyChecker.sol
> Compiling ./contracts/common/test/ReentrancyMock.sol
> Compiling ./contracts/common/test/SignedFixedPointTest.sol
> Compiling ./contracts/common/test/TestableTest.sol
> Compiling ./contracts/common/test/UniswapMock.sol
> Compiling ./contracts/common/test/UnsignedFixedPointTest.sol
> Compiling ./contracts/common/test/WithdrawableTest.sol
> Compiling ./contracts/financial-templates/common/EmergencyShutdownable.sol
> Compiling ./contracts/financial-templates/common/FeePayer.sol
> Compiling ./contracts/financial-templates/common/FundingRateApplier.sol
> Compiling ./contracts/financial-templates/common/SyntheticToken.sol
> Compiling ./contracts/financial-templates/common/TokenFactory.sol
> Compiling ./contracts/financial-templates/common/WETH9.sol
> Compiling ./contracts/financial-templates/common/financial-product-libraries/FinancialProductLibrary.sol
> Compiling ./contracts/financial-templates/common/financial-product-libraries/PreExpirationIdentifierTransformationFinancialProductLibrary.sol
> Compiling ./contracts/financial-templates/common/financial-product-libraries/StructuredNoteFinancialProductLibrary.sol
> Compiling ./contracts/financial-templates/demo/DepositBox.sol
> Compiling ./contracts/financial-templates/expiring-multiparty/ExpiringMultiParty.sol
> Compiling ./contracts/financial-templates/expiring-multiparty/ExpiringMultiPartyCreator.sol
> Compiling ./contracts/financial-templates/expiring-multiparty/ExpiringMultiPartyLib.sol
> Compiling ./contracts/financial-templates/expiring-multiparty/Liquidatable.sol
> Compiling ./contracts/financial-templates/expiring-multiparty/PricelessPositionManager.sol
> Compiling ./contracts/financial-templates/perpetual-multiparty/ConfigStore.sol
> Compiling ./contracts/financial-templates/perpetual-multiparty/ConfigStoreInterface.sol
> Compiling ./contracts/financial-templates/perpetual-multiparty/Perpetual.sol
> Compiling ./contracts/financial-templates/perpetual-multiparty/PerpetualCreator.sol
> Compiling ./contracts/financial-templates/perpetual-multiparty/PerpetualLib.sol
> Compiling ./contracts/financial-templates/perpetual-multiparty/PerpetualLiquidatable.sol
> Compiling ./contracts/financial-templates/perpetual-multiparty/PerpetualPositionManager.sol
> Compiling ./contracts/financial-templates/test/ExpiringMultiPartyMock.sol
> Compiling ./contracts/financial-templates/test/FinancialProductLibraryTest.sol
> Compiling ./contracts/financial-templates/test/FundingRateApplierTest.sol
> Compiling ./contracts/oracle/implementation/Constants.sol
> Compiling ./contracts/oracle/implementation/ContractCreator.sol
> Compiling ./contracts/oracle/implementation/DesignatedVoting.sol
> Compiling ./contracts/oracle/implementation/DesignatedVotingFactory.sol
> Compiling ./contracts/oracle/implementation/FinancialContractsAdmin.sol
> Compiling ./contracts/oracle/implementation/Finder.sol
> Compiling ./contracts/oracle/implementation/Governor.sol
> Compiling ./contracts/oracle/implementation/IdentifierWhitelist.sol
> Compiling ./contracts/oracle/implementation/OptimisticOracle.sol
> Compiling ./contracts/oracle/implementation/Registry.sol
> Compiling ./contracts/oracle/implementation/ResultComputation.sol
> Compiling ./contracts/oracle/implementation/Store.sol
> Compiling ./contracts/oracle/implementation/TokenMigrator.sol
> Compiling ./contracts/oracle/implementation/VoteTiming.sol
> Compiling ./contracts/oracle/implementation/Voting.sol
> Compiling ./contracts/oracle/implementation/VotingToken.sol
> Compiling ./contracts/oracle/implementation/test/GovernorTest.sol
> Compiling ./contracts/oracle/implementation/test/MockAdministratee.sol
> Compiling ./contracts/oracle/implementation/test/OptimisticRequesterTest.sol
> Compiling ./contracts/oracle/implementation/test/ResultComputationTest.sol
> Compiling ./contracts/oracle/implementation/test/VoteTimingTest.sol
> Compiling ./contracts/oracle/implementation/test/VotingTest.sol
> Compiling ./contracts/oracle/interfaces/AdministrateeInterface.sol
> Compiling ./contracts/oracle/interfaces/FinderInterface.sol
> Compiling ./contracts/oracle/interfaces/IdentifierWhitelistInterface.sol
> Compiling ./contracts/oracle/interfaces/OptimisticOracleInterface.sol
> Compiling ./contracts/oracle/interfaces/OracleAncillaryInterface.sol
> Compiling ./contracts/oracle/interfaces/OracleInterface.sol
> Compiling ./contracts/oracle/interfaces/RegistryInterface.sol
> Compiling ./contracts/oracle/interfaces/StoreInterface.sol
> Compiling ./contracts/oracle/interfaces/VotingAncillaryInterface.sol
> Compiling ./contracts/oracle/interfaces/VotingInterface.sol
> Compiling ./contracts/oracle/test/MockOracle.sol
> Compiling ./contracts/oracle/test/MockOracleAncillary.sol
> Compiling ./contracts/oracle/test/VotingAncillaryInterfaceTest.sol
> Compiling ./contracts/oracle/test/VotingInterfaceTest.sol
> Compiling ./contracts/umips/Umip15Upgrader.sol
> Compiling ./contracts/umips/Umip3Upgrader.sol
> Compiling @openzeppelin/contracts/GSN/Context.sol
> Compiling @openzeppelin/contracts/access/Ownable.sol
> Compiling @openzeppelin/contracts/cryptography/ECDSA.sol
> Compiling @openzeppelin/contracts/math/Math.sol
> Compiling @openzeppelin/contracts/math/SafeMath.sol
> Compiling @openzeppelin/contracts/math/SignedSafeMath.sol
> Compiling @openzeppelin/contracts/token/ERC20/ERC20.sol
> Compiling @openzeppelin/contracts/token/ERC20/ERC20Snapshot.sol
> Compiling @openzeppelin/contracts/token/ERC20/IERC20.sol
> Compiling @openzeppelin/contracts/token/ERC20/SafeERC20.sol
> Compiling @openzeppelin/contracts/utils/Address.sol
> Compiling @openzeppelin/contracts/utils/Arrays.sol
> Compiling @openzeppelin/contracts/utils/Counters.sol
> Compiling @openzeppelin/contracts/utils/SafeCast.sol
> Compilation warnings encountered:

    $USER/code/repos/blockchain/eth/UMAprotocol/packages/core/contracts/financial-templates/common/financial-product-libraries/FinancialProductLibrary.sol: Warning: SPDX license identifier not provided in source file. Before publishing, consider adding a comment containing "SPDX-License-Identifier: <SPDX-License>" to each source file. Use "SPDX-License-Identifier: UNLICENSED" for non-open-source code. Please see https://spdx.org for more information.
,$USER/code/repos/blockchain/eth/UMAprotocol/packages/core/contracts/financial-templates/common/financial-product-libraries/PreExpirationIdentifierTransformationFinancialProductLibrary.sol: Warning: SPDX license identifier not provided in source file. Before publishing, consider adding a comment containing "SPDX-License-Identifier: <SPDX-License>" to each source file. Use "SPDX-License-Identifier: UNLICENSED" for non-open-source code. Please see https://spdx.org for more information.
,$USER/code/repos/blockchain/eth/UMAprotocol/packages/core/contracts/financial-templates/common/financial-product-libraries/StructuredNoteFinancialProductLibrary.sol: Warning: SPDX license identifier not provided in source file. Before publishing, consider adding a comment containing "SPDX-License-Identifier: <SPDX-License>" to each source file. Use "SPDX-License-Identifier: UNLICENSED" for non-open-source code. Please see https://spdx.org for more information.
,$USER/code/repos/blockchain/eth/UMAprotocol/packages/core/contracts/financial-templates/test/ExpiringMultiPartyMock.sol: Warning: SPDX license identifier not provided in source file. Before publishing, consider adding a comment containing "SPDX-License-Identifier: <SPDX-License>" to each source file. Use "SPDX-License-Identifier: UNLICENSED" for non-open-source code. Please see https://spdx.org for more information.
,$USER/code/repos/blockchain/eth/UMAprotocol/packages/core/contracts/financial-templates/test/FinancialProductLibraryTest.sol: Warning: SPDX license identifier not provided in source file. Before publishing, consider adding a comment containing "SPDX-License-Identifier: <SPDX-License>" to each source file. Use "SPDX-License-Identifier: UNLICENSED" for non-open-source code. Please see https://spdx.org for more information.
,$USER/code/repos/blockchain/eth/UMAprotocol/packages/core/contracts/common/test/BalancerMock.sol:13:34: Warning: Unused function parameter. Remove or comment out the variable name to silence this warning.
    function getSpotPriceSansFee(address tokenIn, address tokenOut)
                                 ^-------------^
,$USER/code/repos/blockchain/eth/UMAprotocol/packages/core/contracts/common/test/BalancerMock.sol:13:51: Warning: Unused function parameter. Remove or comment out the variable name to silence this warning.
    function getSpotPriceSansFee(address tokenIn, address tokenOut)
                                                  ^--------------^
,$USER/code/repos/blockchain/eth/UMAprotocol/packages/core/contracts/common/test/OneSplitMock.sol:31:9: Warning: Unused function parameter. Remove or comment out the variable name to silence this warning.
        uint256 parts,
        ^-----------^
,$USER/code/repos/blockchain/eth/UMAprotocol/packages/core/contracts/common/test/OneSplitMock.sol:32:9: Warning: Unused function parameter. Remove or comment out the variable name to silence this warning.
        uint256 flags // See constants in IOneSplit.sol
        ^-----------^
,$USER/code/repos/blockchain/eth/UMAprotocol/packages/core/contracts/common/test/OneSplitMock.sol:44:9: Warning: Unused function parameter. Remove or comment out the variable name to silence this warning.
        uint256[] memory distribution,
        ^---------------------------^
,$USER/code/repos/blockchain/eth/UMAprotocol/packages/core/contracts/common/test/OneSplitMock.sol:45:9: Warning: Unused function parameter. Remove or comment out the variable name to silence this warning.
        uint256 flags
        ^-----------^
,$USER/code/repos/blockchain/eth/UMAprotocol/packages/core/contracts/common/test/OneSplitMock.sol:46:40: Warning: Unused function parameter. Remove or comment out the variable name to silence this warning.
    ) public payable override returns (uint256 returnAmount) {
                                       ^------------------^
,$USER/code/repos/blockchain/eth/UMAprotocol/packages/core/contracts/financial-templates/common/financial-product-libraries/FinancialProductLibrary.sol:23:69: Warning: Unused function parameter. Remove or comment out the variable name to silence this warning.
    function transformPrice(FixedPoint.Unsigned memory oraclePrice, uint256 requestTime)
                                                                    ^-----------------^
,$USER/code/repos/blockchain/eth/UMAprotocol/packages/core/contracts/financial-templates/common/financial-product-libraries/FinancialProductLibrary.sol:39:9: Warning: Unused function parameter. Remove or comment out the variable name to silence this warning.
        FixedPoint.Unsigned memory oraclePrice,
        ^------------------------------------^
,$USER/code/repos/blockchain/eth/UMAprotocol/packages/core/contracts/financial-templates/common/financial-product-libraries/FinancialProductLibrary.sol:51:64: Warning: Unused function parameter. Remove or comment out the variable name to silence this warning.
    function transformPriceIdentifier(bytes32 priceIdentifier, uint256 requestTime)
                                                               ^-----------------^
,$USER/code/repos/blockchain/eth/UMAprotocol/packages/core/contracts/financial-templates/test/FinancialProductLibraryTest.sol:28:69: Warning: Unused function parameter. Remove or comment out the variable name to silence this warning.
    function transformPrice(FixedPoint.Unsigned memory oraclePrice, uint256 requestTime)
                                                                    ^-----------------^
,$USER/code/repos/blockchain/eth/UMAprotocol/packages/core/contracts/financial-templates/test/FinancialProductLibraryTest.sol:40:9: Warning: Unused function parameter. Remove or comment out the variable name to silence this warning.
        FixedPoint.Unsigned memory price,
        ^------------------------------^
,$USER/code/repos/blockchain/eth/UMAprotocol/packages/core/contracts/financial-templates/test/FinancialProductLibraryTest.sol:48:39: Warning: Unused function parameter. Remove or comment out the variable name to silence this warning.
    function transformPriceIdentifier(bytes32 priceIdentifier, uint256 requestTime)
                                      ^---------------------^
…