@mrice32 2021-02-01
Closing due to inactivity. We will likely do this at some point in the future, and I still think this is a great idea, but I think it makes sense to open a new PR when we do our next round of audits/contract upgrades!